Supports batches that have both read and write ops. (#98)
The previous implementation set up batches but delayed setting the
isrdmsr field until batch execution, setting it depending on
whether or not the executing call was read_batch() or
write_batch(). This is not necessarily a poor design choice, but
it is not obvious which batches are intented to be read only and
which are r/w. Nor does the previous design permit batches that,
say, read and store the previous values of the MSRs before writing
new values.

The implementation contained in this patch forces the user to
declare whether batch operations are read or write when the batch
is initialized. Batches that could formerly be read or written
are now replaced by two distinct batches, one for reading and the
second for writing (e.g., the COUNTERS_CTRL batch has been
replaced by RD_COUNTERS_CTRL and WR_COUNTERS_CTRL).

This patch adds a parameter (int isrdmsr) to load_*_batch.  This
is an intrusive change, as all of the *_features.[ch] code made
use of this API.

This patch also replaces the separate read_batch() and write_batch()
calls with execute_batch().