Allow travis script to work with older msr-safe. (#137)
load-msr-safe.sh had not yet been convered to use the allowlist
terminology present in msr-safe 1.4 and greater.  This PR pulls
the current msr-safe version out of /sys/module/msr_safe/version
to determine if whitelist or allowlist should be used.

This PR introduces a dependence on coreutils 7 or greater as that
is the version where sort -V was introduced.

* Convert whitelist to allowlist throughout code and docs
* This PR introduces a dependence on the python packaging module.
It can be installed in Ubuntu via `apt-get install python3-packaging`.
* Changed example to reflect output of msr-safe v1.4.
* Changed "python" to "python3" to allow execution on Ubuntu
systems.
* Modified code so that USE_MSR_SAFE_BEFORE_1_5_0 has to be
defined to test with earlier versions of msr-safe.  How that
defined value shows up is left as an exercise for the reader.