add new CMake option for toggling between msr-safe versions (#148)
- update quartz and catalyst host configs with new option
- add cmake output for this option
- small cleanup of other outputs in cmake output

Version 1.5.0 and later of msr-safe uses the msr_allowlist interface, while
versions prior to 1.5.0 use the msr_whitelist interface. Variorum encounters
both versions of msr-safe on different machines, so this provides an easy way
of switching between the two.

The default msr-safe version in variorum's build will be "before v1.5.0" (i.e.,
msr_whitelist interface) until more systems start migrating to the newer
msr-safe version. Set this option to ON in host config files for quartz and
catalyst for redundancy.